Move impl Evaluate for Constraint
