+++ rust/ommx/src/evaluate.rs
@@ -1,9 +1,8 @@
 use crate::v1::{
-    Constraint, Equality, EvaluatedConstraint, Function, Instance, Optimality, Relaxation,
-    RemovedConstraint, SampleSet, SampledConstraint, SampledDecisionVariable, Samples, Solution,
-    State,
+    Function, Instance, Optimality, Relaxation, SampleSet, SampledDecisionVariable, Samples,
+    Solution, State,
 };
-use anyhow::{bail, Context, Result};
+use anyhow::{bail, Result};
 use std::collections::{hash_map::Entry as HashMapEntry, BTreeSet, HashMap};
 
 /// Evaluate with a [State]
@@ -22,120 +21,6 @@ pub trait Evaluate {
 
     /// Decision variable IDs required for evaluation
     fn required_ids(&self) -> BTreeSet<u64>;
-}
-
-impl Evaluate for Constraint {
-    type Output = EvaluatedConstraint;
-    type SampledOutput = SampledConstraint;
-
-    fn evaluate(&self, solution: &State) -> Result<Self::Output> {
-        let evaluated_value = self.function().evaluate(solution)?;
-        let used_decision_variable_ids = self
-            .function()
-            .used_decision_variable_ids()
-            .into_iter()
-            .collect();
-        Ok(EvaluatedConstraint {
-            id: self.id,
-            equality: self.equality,
-            evaluated_value,
-            used_decision_variable_ids,
-            name: self.name.clone(),
-            subscripts: self.subscripts.clone(),
-            parameters: self.parameters.clone(),
-            description: self.description.clone(),
-            dual_variable: None,
-            removed_reason: None,
-            removed_reason_parameters: Default::default(),
-        })
-    }
-
-    fn partial_evaluate(&mut self, state: &State) -> Result<()> {
-        let Some(f) = self.function.as_mut() else {
-            // Since empty function means zero constant, we can return an empty set
-            return Ok(());
-        };
-        f.partial_evaluate(state)
-    }
-
-    fn evaluate_samples(&self, samples: &Samples) -> Result<Self::SampledOutput> {
-        let evaluated_values = self.function().evaluate_samples(samples)?;
-        let feasible: HashMap<u64, bool> = evaluated_values
-            .iter()
-            .map(|(sample_id, value)| {
-                if self.equality() == Equality::EqualToZero {
-                    return Ok((*sample_id, value.abs() < 1e-6));
-                }
-                if self.equality() == Equality::LessThanOrEqualToZero {
-                    return Ok((*sample_id, *value < 1e-6));
-                }
-                bail!("Unsupported equality: {:?}", self.equality());
-            })
-            .collect::<Result<_>>()?;
-        Ok(SampledConstraint {
-            id: self.id,
-            evaluated_values: Some(evaluated_values),
-            used_decision_variable_ids: self
-                .function()
-                .used_decision_variable_ids()
-                .into_iter()
-                .collect(),
-            name: self.name.clone(),
-            subscripts: self.subscripts.clone(),
-            parameters: self.parameters.clone(),
-            description: self.description.clone(),
-            equality: self.equality,
-            feasible,
-            removed_reason: None,
-            removed_reason_parameters: Default::default(),
-        })
-    }
-
-    fn required_ids(&self) -> BTreeSet<u64> {
-        self.function
-            .as_ref()
-            .map_or(BTreeSet::new(), |f| f.used_decision_variable_ids())
-    }
-}
-
-impl Evaluate for RemovedConstraint {
-    type Output = EvaluatedConstraint;
-    type SampledOutput = SampledConstraint;
-
-    fn evaluate(&self, solution: &State) -> Result<Self::Output> {
-        let mut out = self
-            .constraint
-            .as_ref()
-            .context("RemovedConstraint does not contain constraint")?
-            .evaluate(solution)?;
-        out.removed_reason = Some(self.removed_reason.clone());
-        out.removed_reason_parameters = self.removed_reason_parameters.clone();
-        Ok(out)
-    }
-
-    fn partial_evaluate(&mut self, state: &State) -> Result<()> {
-        self.constraint
-            .as_mut()
-            .context("RemovedConstraint does not contain constraint")?
-            .partial_evaluate(state)
-    }
-
-    fn evaluate_samples(&self, samples: &Samples) -> Result<Self::SampledOutput> {
-        let mut evaluated = self
-            .constraint
-            .as_ref()
-            .expect("RemovedConstraint does not contain constraint")
-            .evaluate_samples(samples)?;
-        evaluated.removed_reason = Some(self.removed_reason.clone());
-        evaluated.removed_reason_parameters = self.removed_reason_parameters.clone();
-        Ok(evaluated)
-    }
-
-    fn required_ids(&self) -> BTreeSet<u64> {
-        self.constraint
-            .as_ref()
-            .map_or(BTreeSet::new(), |c| c.required_ids())
-    }
 }
 
 impl Evaluate for Instance {
